requirements.txt
# Core dependencies
PyYAML>=5.4                # for reading config files
SpeechRecognition>=3.8.1   # high-level STT interface (Google, Sphinx)
vosk>=0.3.45               # offline speech recognition
whisper-client>=0.1.0      # if using OpenAI Whisper locally
pyttsx3>=2.90              # offline TTS engine
Coqui-TTS>=0.22.1          # neural TTS (optional)
spacy>=3.5.0               # NLP toolkit for intent/entity parsing
apscheduler>=3.9.1         # scheduling reminders/alarms
dateparser>=1.1.0        # parse natural language dates (optional)
sqlalchemy>=1.4            # ORM layer for SQLite
faiss-cpu>=1.7.3           # vector embeddings store (or Chroma if preferred)
sentence-transformers>=2.2.2  # embeddings model for vector store
numpy>=1.21.0              # numerical computing for ML components
stable-baselines3>=2.1.0   # reinforcement learning algorithms
networkx>=3.1              # optional: knowledge graph
requests>=2.28.0           # web queries (weather, news)
beautifulsoup4>=4.12.0     # simple scraping if needed
phue==1.1                  # Philips Hue integration (smart home)
pywemo>=0.6.0              # Belkin WeMo integration
pytest>=7.2.0              # for running unit tests
python-dotenv>=0.21.0      # optionally load environment variables
openai>=1.0.0             # ChatGPT integration for conversation
rasa>=3.1                 # optional: use Rasa NLU engine
pydantic>=1.10            # configuration validation
redis>=4.5.0              # optional Redis memory backend
